Workaround in rapidjson/reader.h for NVCC (#237)
* Workaround in rapidjson/reader.h for NVCC

- bug in CUDA 11.5

* Miscellaneous fixes

- fix for gotcha + nvcc cuda 11.5 fix
- TIMEMORY_CUDA_ARCH define
- Better handling of TIMEMORY_USE_CUDA_HALF
- minor fix to timemory-ert executable output folder
- Updated nccl-tests folder
- Python set_write_metadata
- TIMEMORY_BUILD_PYTHON_HIDDEN option (advanced)

* CUPTI PCSampling fix

+++ source/timemory/components/cupti/cupti_pcsampling.hpp
@@ -101,7 +101,7 @@ struct cupti_pcsampling
 
     auto get_laps() const { return value.totalSamples; }
 
-    std::vector<uint32_t>           get() const;
+    std::vector<int64_t>            get() const;
     static std::vector<std::string> label_array();
 
     static void cleanup() { cupti::pcstall::allocate_arrays(0); }
